% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_cedar_odbc.R
\name{connect_cedar_odbc}
\alias{connect_cedar_odbc}
\title{Make a database connection to CEDAR}
\usage{
connect_cedar_odbc(read_only = TRUE)
}
\arguments{
\item{read_only}{Create a READ ONLY connection. Default = TRUE}
}
\value{
a valid SQL Server connection object to the Reporting database
}
\description{
Connect to CEDAR using the \code{odbc} package
Requires:
\itemize{
\item User configuration file: usr.dat.
\item package DBI   For basic database access
\item package odbc  For access to SQL Server databases
}
}
\examples{
\dontrun{
CEDAR <- connect_cedar_odbc()
}

}
\seealso{
\link{extract_cedar} and \link{extract_cedar_linelist}
}
