% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cedar.R
\name{extract_cedar}
\alias{extract_cedar}
\title{Extract data from CEDAR}
\usage{
extract_cedar(
  sql =
    "\\nSELECT   TOP 10   SnapshotDate, NotificationId, OnsetDate  as EpisodeDate\\n     , SymptomOnsetDate  as OnsetDate, NotificationDate\\n     , SexName  as Sex, IndigGroupName as IndigGroup, AgeGroupCode  as AgeGroup\\n     , HHSName as HHS\\n     , DiseaseName as Disease, Pathogen, Level1, Level2, Level3\\n  FROM nd.NotificationsView n \\n WHERE DiseaseName = 'Influenza (lab confirmed)'\\n  AND YEAR(OnsetDate) = YEAR(SnapshotDate)\\n ORDER BY OnsetDate, NotificationId\\n                          "
)
}
\arguments{
\item{sql}{A valid SQL command to extract data from CEDAR}
}
\value{
a tibble dataframe
}
\description{
Extract data from CEDAR using the supplied SQL using package \code{odbc}
and a connection object from the \code{rphido} function \code{connect_cedar_odbc()}
}
\examples{
\dontrun{
sql <= "SELECT HHSName from nd.HHS where HHSName like 'Metro\%'"
metro_hhs <- extract_cedar(sql)
}
# Uses the function `connect_cedar_odbc()` to create a connection to CEDAR

}
\seealso{
\link{extract_cedar_linelist} and \link{connect_cedar_odbc}
}
