% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cedar_linelist.R
\name{extract_cedar_linelist}
\alias{extract_cedar_linelist}
\title{Extract a line list of notifications data from CEDAR}
\usage{
extract_cedar_linelist(
  disease = "Measles",
  ytd = TRUE,
  start_date,
  end_date = "2099-12-31",
  filter_onset = TRUE,
  select_extra = "",
  filter_extra = "",
  sql_only = FALSE
)
}
\arguments{
\item{disease}{The disease to use for this query. Use the \strong{exact spelling} from the Online Weekly Report. Defaults to \code{"Measles"}}

\item{ytd}{Use the current year as the data range for the query. Overrides \code{start_date} and \code{end_date}. Defaults to \code{TRUE}}

\item{start_date}{The start_date for the date_range for the query.}

\item{end_date}{The end_date for the date_range for the query. Defaults to \code{"2099-12-31"}.}

\item{filter_onset}{Use OnsetDate rather than NotificationDate for the date filter. Defaults to \code{TRUE}.}

\item{select_extra}{Additional fields to extract. For more than one field include a comma separated character string. Defaults to \code{""}.}

\item{filter_extra}{Additional filter to apply to the extract. Use valid SQL format such as \code{"SexName = 'Male'"}. The first \code{ AND} will be added automatically. For more than one filter include additional \code{ AND} in the character string.}

\item{sql_only}{return a character string of the custom SQL. The data is NOT extracted.}
}
\value{
a tibble dataframe with the requested data or a character string of the custom sql if \code{sql_only = TRUE}
}
\description{
Extract a custom linelist of notifications data from CEDAR.
An SQL statement is created by the function. The user can customise the SQL query by choosing parameters.
The query is based on  the View \code{nd.NotificationsView} and includes common fields by default.
}
\details{
\itemize{
\item Choose the \code{disease} for the extract.
\item By default the current year is used as the date range, \code{ytd = TRUE}.
If \code{ytd = FALSE} then \code{start_date} and \code{end_date} are used as the date_range.
\item By default the extract is filtered by \code{OnsetDate}. To filter by \code{NotificationDate}, use \code{filter_onset = FALSE}.
\item Add more fields from \code{nd.NotificationsView} by using the \code{select_extra} parameter.
\item Add more filters by using the \code{filter_extra} parameter.
}

#######- The function will output the custom SQL to the screen and extract the data with \code{qcider::extract_RD(sql)}.
\itemize{
\item The function will output the custom SQL to the screen and extract the data with \code{rohido::extract_cedar(sql)}.
By choosing \code{sql_only = TRUE} the function will return a character string of the SQL command rather than the data.
\item If more flexibility is required then use \code{extract_from_RD} and supply your own SQL.
}
}
\examples{
\dontrun{
# using all defaults: disease = "Measles" ytd = TRUE.
extract_cedar_linelist()

# to choose disease is Influenza
extract_cedar_linelist(disease = "Influenza (lab confirmed)")

# custom start_date and end_date
 extract_cedar_linelist(disease = "Syphilis (infectious)", start_date = "2017-01-01"
                    , end_date = "2018-06-30")

# filter on NotificationDate for Tuberculosis
 extract_cedar_linelist(disease = "Tuberculosis", filter_onset = FALSE)

# add an additional field: LGAName
extract_cedar_linelist(disease = "Dengue", select_extra = "LGAName")

# add an addditional filter: males only
  # note use of the exact name from nd.NotificationsView `SexName` and single quotes around Male
extract_RD_linelist(disease = "Chlamydia (STI)", filter_extra = "SexName = 'Male'")

# add two addditional filter: males only and age at onset over 20
  # note use of the exact name from nd.NotificationsView `SexName` and single quotes around Male
  # and additional AND clause in the character string for the filter
extract_RD_linelist(disease = "Chlamydia (STI)"
    , filter_extra = "SexName = 'Male' AND AgeAtOnsetYears > 20")
}
}
\seealso{
\link{extract_cedar} \link{connect_cedar_odbc}
}
