% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_cedar_names.R
\name{show_cedar_names}
\alias{show_cedar_names}
\title{Show field names of notification in CEDAR.}
\usage{
show_cedar_names()
}
\value{
a character vector of the field names of notifications in CEDAR.
}
\description{
Return a vector of names of the fields of the main Notifications view
\code{nd.NotificationsView} in CEDAR.
These field names can can be used in the custom sql command for the
\code{extract_from_cedar} function,
\verb{and in the }select_extra\verb{parameter of}extract_cedar_linelist`.
}
\examples{
\dontrun{
show_cedar_names()
cedar_nv_names <- show_cedar_names()
}
}
\seealso{
\link{extract_cedar_linelist} and \link{extract_cedar}  and \link{connect_cedar}
}
