% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_rip.R
\name{str_rip}
\alias{str_rip}
\title{Rip (split) a string into a vector by a simple string separator}
\usage{
str_rip(string, pattern = ",")
}
\arguments{
\item{string}{A string to be split}

\item{pattern}{A separator pattern used to split the string.
Note: this a simple string, NOT a regular expression}
}
\value{
a character matrix with columns for each split of the input string
and a row for each input string
}
\description{
Split a string \code{string} (or vector of strings) by a simple separator.
The separator \code{pattern} is a simple string, NOT a regular expression.
This is a simple wrapper function for:
\code{stringr::str_split(string, stringr::fixed(pattern), simplify = TRUE)}
}
\examples{
str_rip("x,y,z")  # defaults to comma as a separator pattern.

test_chars <- tibble::tribble(
  ~pipe, ~comma, ~space, ~dash, ~slash, ~backslash, ~FRED, ~underscore,
  "a|b|c", "a,b,c", "a b c", "a-b-c", "a/b/c", "a\\\\b\\\\c", "aFREDbFREDc", "a_b_c",
  "d|e|f", "d,e,f", "d e f", "d-e-f", "d/e/f", "d\\\\e\\\\f", "dFREDeFREDf", "d_e_f"
)

str_rip(test_chars$pipe, "|")
str_rip(test_chars$comma, ",")
str_rip(test_chars$comma)
str_rip(test_chars$space, " ")
str_rip(test_chars$dash, "-")
str_rip(test_chars$slash, "/")
str_rip(test_chars$backslash, "\\\\")
str_rip(test_chars$FRED, "FRED")
str_rip(test_chars$underscore, "_")
str_rip(test_chars$dash, "FAIL")
}
